requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
distro==1.9.0
docx2txt==0.8
emoji==2.10.0
environs==9.5.0
exceptiongroup==1.2.0
fastapi==0.109.0
filetype==1.2.0
frozenlist==1.4.1
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
langchain==0.1.3
langchain-community==0.0.15
langchain-core==0.1.15
langchain-openai==0.0.3
langdetect==1.0.9
langsmith==0.0.83
lxml==5.1.0
marshmallow==3.20.2
minio==7.2.3
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.26.3
openai==1.9.0
packaging==23.2
pandas==2.2.0
protobuf==4.25.2
pyarrow==15.0.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.5.3
pydantic_core==2.14.6
pymilvus==2.3.6
pypdf==4.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.1.2
python-magic==0.4.27
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.6.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
starlette==0.35.1
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
unstructured==0.12.2
unstructured-client==0.15.2
uvicorn==0.27.0
wrapt==1.16.0
yarl==1.9.4
# urllib3==1.26.6
